fix(CHYT): default pool tree should be loaded properly [YTFRONT-3683]
