+++ packages/ui/src/ui/store/selectors/global/pool-trees.ts
@@ -1,0 +1,3 @@
+import {RootState} from '../../../store/reducers';
+
+export const getGlobalDefaultPoolTreeName = (state: RootState) => state.global.poolTreeDefault;
